requirements.txt
requests==2.31.0
python-dotenv==1.0.0
rich==13.5.2
pydantic==2.4.2
typer==0.9.0
streamlit==1.32.0 